requirements.txt
Django
celery
defusedxml
kombu
django-dirtyfields
djangorestframework
